+++ .gitignore
@@ -38,3 +38,5 @@ out/
 
 ### Mac OS ###
 .DS_Store
+
+application.properties
