+++ src/main/java/gift/controller/kakao/KakaoController.java
@@ -20,52 +20,16 @@ import org.springframework.web.client.RestClient;
 @RequestMapping("/")
 public class KakaoController {
 
-    private final KakaoProperties kakaoProperties;
-    private final RestClient restClient;
-    private final ObjectMapper objectMapper;
     private final KakaoService kakaoService;
 
-    public KakaoController(KakaoProperties kakaoProperties, KakaoService kakaoService) {
-        this.kakaoProperties = kakaoProperties;
-        this.restClient = RestClient.builder().build();
-        this.objectMapper = new ObjectMapper();
+    public KakaoController(KakaoService kakaoService) {
         this.kakaoService = kakaoService;
     }
 
     @GetMapping
     public ResponseEntity<String> kakaoLogin(@RequestParam("code") String authCode) {
-        String tokenRequestUrl = "https://kauth.kakao.com/oauth/token";
-
-        MultiValueMap<String, String> body = createRequestBody(authCode);
-        String tokenResponse = sendTokenRequest(tokenRequestUrl, body);
-        String accessToken = extractAccessToken(tokenResponse);
+        String accessToken = kakaoService.sendTokenRequest(authCode);
         return ResponseEntity.status(HttpStatus.OK).body(accessToken);
     }
 
-    private MultiValueMap<String, String> createRequestBody(String authCode) {
-        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("grant_type", "authorization_code");
-        body.add("client_id", kakaoProperties.clientId());
-        body.add("redirect_uri", kakaoProperties.redirectUri());
-        body.add("code", authCode);
-        return body;
-    }
-
-    private String sendTokenRequest(String url, MultiValueMap<String, String> body) {
-        return restClient.post()
-                        .uri(URI.create(url))
-                        .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                        .body(body)
-                        .retrieve()
-                        .body(String.class);
-    }
-
-    private String extractAccessToken(String tokenResponse) {
-        try {
-            JsonNode jsonNode = objectMapper.readTree(tokenResponse);
-            return jsonNode.path("access_token").asText();
-        } catch (IOException e) {
-            throw new RuntimeException("Failed to parse access token", e);
-        }
-    }
 }
